Update SteamedHams.py
Fixed the Loot verb to work properly and not spam the output with item not found or container not found messages when containers or items are correctly used. The previous version was a temporary measure and it iterated over all containers and items and would give not found errors for all non matching items.