more groundwork
more groundwork before the game is feature complete. Some of the commands are going to be a total pain because to refactor them I am going to have to write the whole thing from scratch because the backend is different. Hopefully it will all be worth it in the end.