Take from container added
Still need to round out the details but the take from container is working. I will need to add a check for the is container bool and also get rid of the deprecated container class. It worked great in the original python script but here we are using a lot of arrays so mixing classes just does not work.