added place command.
with this we are very close to having the game up and running. Still need to implement Talk and Use as well as the H.A.M.S script but once that's done I can rebuild all the game paths and logic and the MVP will be ready for release.