Tied kitchen fire into lunch scene
I hate stack overflow, I had an issue with the loop still triggering the last else if even if I did break or return. And I did make sure it was not calling the method and thus starting the loop more than once but people on stack overflow cannot freaking read. or just choose not to. anyway long story short I implemented this workaround and it seems stable.